Fix answer display in student workbook
Fixed the answer display in the student workbook.  Old version just "hid" answers in the student version by outputting them in white text, which allowed students using the PDF version of the book to view them by highlighting the text.  New version uses the environ package to suppress output of the solution environment completely when in the student display.