Fix answer display & propagate to all activities on branch
Made further changes to the solution environment. In particular:
- information to display in student handbook is now a required input for the environment
- \studentdisplay{} and \instructordisplay{} macros are no longer used and will cause errors
These changes simplified solution input and solved a number of problems with display of links and images that were requiring \instructordisplay{} to be used as a workaround.

Also propagated the new format to all activities on the current branch, and updated the activity template.  Changes will still need to be propagated to all new activities that haven't been merged into this branch yet.